pyproject.toml
[project]
name = "mlops-lab-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
]
